replace FroseAiOptimizer with FedAvg and restructure optimizer module
