+++ froseai/__init__.py
@@ -1,13 +1,13 @@
 from .server import FroseAiServer
 from .aggregator import FedAvgAggregator
-from .flow import FroseAiOptimizer
+from .optimizer import FedAvg
 from .datasets import FedDatasetsClassification
 from .context import FroseArguments
 
 __all__ = [
     "FroseAiServer",
     "FedAvgAggregator",
-    "FroseAiOptimizer",
+    "FedAvg",
     "FedDatasetsClassification",
     "FroseArguments",
 ]
