--- froseai/flow/__init__.py
@@ -1,1 +1,0 @@
-from .optimizer import FedAvg as FroseAiOptimizer
