+++ froseai/optimizer/__init__.py
@@ -1,0 +1,5 @@
+from .fedavg import FedAvg
+
+__all__ = [
+    "FedAvg"
+]
